Move FluidTransferUtils
