+++ compat/transfer-api-compat/src/main/kotlin/xyz/bluspring/kilt/compat/transfer/fluid/FabricFluidStorageCapability.kt
@@ -5,8 +5,8 @@ import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant
 import net.fabricmc.fabric.api.transfer.v1.storage.Storage
 import net.neoforged.neoforge.fluids.FluidStack
 import net.neoforged.neoforge.fluids.capability.IFluidHandler
-import xyz.bluspring.kilt.compat.transfer.fluid.FluidTransferUtils.toDroplets
-import xyz.bluspring.kilt.compat.transfer.fluid.FluidTransferUtils.toMillibuckets
+import xyz.bluspring.kilt.util.neoforge.fluid.FluidTransferUtils.toDroplets
+import xyz.bluspring.kilt.util.neoforge.fluid.FluidTransferUtils.toMillibuckets
 
 open class FabricFluidStorageCapability(val storage: Storage<FluidVariant>) : IFluidHandler {
     override fun getTanks(): Int {
