+++ src/main/kotlin/xyz/bluspring/kilt/util/neoforge/fluid/FluidTransferUtils.kt
@@ -1,20 +1,24 @@
-package xyz.bluspring.kilt.compat.transfer.fluid
+package xyz.bluspring.kilt.util.neoforge.fluid
 
 import io.github.fabricators_of_create.porting_lib.transfer.TransferUtil
 
 object FluidTransferUtils {
+    @JvmStatic
     fun Long.toMillibuckets(): Int {
         return TransferUtil.truncateLong(this / 81)
     }
 
+    @JvmStatic
     fun Long.toMillibucketsLong(): Long {
         return this / 81
     }
 
+    @JvmStatic
     fun Int.toDroplets(): Long {
         return this * 81L
     }
 
+    @JvmStatic
     fun Long.toDroplets(): Long {
         return this * 81L
     }
